Fix My link for adding an add-on repository not doing anything
The supervisor_add_addon_repository My redirect targets
/config/apps/available, but the repository_url handling was removed
from that page in #29931 when repository management moved to a
dedicated page, so the link opened the store and silently ignored the
parameter.

Restore the handling on the store page, now with a confirmation dialog
before adding the repository (matching the missing-repository flow on
the app dashboard) and an already-added check. Also remove the
unconfirmed auto-add on the repositories page, which nothing links to
anymore.
